refactor: ImageWorker -> ImageManager
